added "custom base classes" support
It is now possible to specify custom base classes for wxGlade-generated
ones. At the moment, it works only for the Python code generator, but I
plan to implement it also for C++.

This feature is not really meant to be used unless the "overwrite
existing sources" option  is set: if not, the generated code might be
incorrect.