Implemented "exraproperties" property.
This allows you to add more properties to the generated code for the
widgets.
For each pair (name, value) in extraproperties for a widget, a line

widget.SetName(value)

will be added in __set_properties. That is, the name of the property is
capitalized, and the value is taken as it is. This works for Python, C++
and Perl. For XRC, the extra line would be

<name>value</name>

(notice that there is no capitalization).