Add support for widget builder that don't support all wx versions
Some new widgets like wxHyperlinkCtrl are not available for wxWidgets
2.6. Thereby a new class attribute called "supported_by" is introducted.

The attribute has set in the code generator class are well as in the
class to edit the widget.

The properties window shows a note at bottom if the widget is not
supported at all wx versions supported by wxGlade.