Rework __do_layout() and __set_properties() in all code generators
Now the functions __do_layout() and __set_properties() are generated by
using the same code. This causes some minor changes in the code
generated by wxGlade for both functions for Perl and Lisp.

Both changes effects only Perl and Lisp code generated for both
functions and if in-place editing is used.

For Perl:
Some empty lines are gone.

For Lisp:
The round brackets are now outside of the blocks "begin wxGlade" and
"end wxGlade" in the Lisp code now.