Rework generating constructor code in all code generators
This causes some minor changes in the code generated by wxGlade for
Lisp, Perl, Python and C++.

For Perl:
Some empty lines are gone.

For Lisp:
The round brackets are now outside of the blocks "begin wxGlade" and
"end wxGlade" in the Lisp code now. There is an additional new line in
front of the dependencies now.

For Python:
A single quotation mark has changed in a printed message.

For C++:
Event table and event handler stubs have been moved later in the same
file. This change was necessary to equalise parts of the C++ code
generator code.