Fix a infinite loop within sizing a widget (sf bug #170)
There are an infinite loop of wxSizeEvents. All events have the same id.
It looks currently like a bug in the underlaying wx libraries especially
in the GTK part. The bug doesn't occur on Windows.

The issue probably occur only within EditGrid.

This is workaround prevents the propagation if the size hasn't changed.