Fix PyDeadObject errors and crashes during cut and paste
During pasting a widget into a free position the SizerSlot representing
this position will be deleted and a new widget will be placed at this
position.

If the "paste" action is triggered via SizerSlot's PopupMenu, the
SizerSlot's internally used widgets e.g. the mentioned PopupMenu are
deleted even if they are still in use. This logical bug caused errors.

Now the execution of the "paste" action is moved away from the context
of the PopupMenu forward to the event handling mechanism of wx. Thereby
the objects to delete will not be in use anymore.

Steps to reproduce:
 1. open a new frame and place a button in to
 2. Cut the button (CTRL-X or via context menu)
 3. Paste the button again (via context menu)
(grafted from 36aa7dc90d02ddcee816f806ef7b4240d74f3c60)