update version information and make shell script more reliable
When wxGlade is installed with the --user parameter to setup.py, its
files go into ~/.local . The wxglade shell script should also look for
the whole path to the ``egg'' directory.

In addition to this, the wxglade shell script should not stick to the
first Python version it finds, as it may not be the appropriate one for
wxGlade.